+++ src/AmyrisBio/AssemblyInfo.fs
@@ -4,10 +4,10 @@ open System.Reflection
 [<assembly: AssemblyTitleAttribute("AmyrisBio")>]
 [<assembly: AssemblyProductAttribute("AmyrisBio")>]
 [<assembly: AssemblyDescriptionAttribute("Amyris FSharp BioLib")>]
-[<assembly: AssemblyVersionAttribute("2.0.20")>]
-[<assembly: AssemblyFileVersionAttribute("2.0.20")>]
+[<assembly: AssemblyVersionAttribute("2.0.21")>]
+[<assembly: AssemblyFileVersionAttribute("2.0.21")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "2.0.20"
-    let [<Literal>] InformationalVersion = "2.0.20"
+    let [<Literal>] Version = "2.0.21"
+    let [<Literal>] InformationalVersion = "2.0.21"